Moving to utils.h
