+++ test/parallel_api/iterator/input_data_sweep_usm_allocator.pass.cpp
@@ -26,37 +26,6 @@
 // are being tested, and to limit the number of types to be within reason.
 
 #if TEST_DPCPP_BACKEND_PRESENT
-namespace TestUtils
-{
-template <typename Iter, typename ValueType = typename std::iterator_traits<Iter>::value_type>
-using __default_alloc_vec_iter = typename std::vector<ValueType>::iterator;
-
-template <typename Iter, typename ValueType = typename std::iterator_traits<Iter>::value_type>
-using __usm_shared_alloc_vec_iter =
-    typename std::vector<ValueType, typename sycl::usm_allocator<ValueType, sycl::usm::alloc::shared>>::iterator;
-
-template <typename Iter, typename ValueType = typename std::iterator_traits<Iter>::value_type>
-using __usm_host_alloc_vec_iter =
-    typename std::vector<ValueType, typename sycl::usm_allocator<ValueType, sycl::usm::alloc::host>>::iterator;
-
-// Evaluates to true if the provided type is an iterator with a value_type and if the implementation of a
-// std::vector<value_type, Alloc>::iterator can be distinguished between three different allocators, the
-// default, usm_shared, and usm_host. If all are distinct, it is very unlikely any non-usm based allocator
-// could be confused with a usm allocator.
-template <typename Iter, typename Void = void>
-struct __vector_impl_distinguishes_usm_allocator_from_default : std::false_type
-{
-};
-
-template <typename Iter>
-struct __vector_impl_distinguishes_usm_allocator_from_default<
-    Iter, std::enable_if_t<!std::is_same_v<__default_alloc_vec_iter<Iter>, __usm_shared_alloc_vec_iter<Iter>> &&
-                           !std::is_same_v<__default_alloc_vec_iter<Iter>, __usm_host_alloc_vec_iter<Iter>> &&
-                           !std::is_same_v<__usm_host_alloc_vec_iter<Iter>, __usm_shared_alloc_vec_iter<Iter>>>>
-    : std::true_type
-{
-};
-} //namespace TestUtils
 
 template <typename T, int __recurse, typename Policy>
 void
